+++ packages/canvas-tokens-docs/.storybook/Label.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+
+const flexStyles = {
+  display: 'flex',
+  alignItems: 'center',
+  gap: '8px',
+  width: '100%',
+  justifyContent: 'space-between',
+  paddingInlineEnd: '16px',
+};
+
+const deprecatedBoxStyles = {
+  padding: '0 8px',
+  backgroundColor: 'rgba(219, 225, 233, 1)',
+  color: 'rgba(93, 103, 117, 1)1)',
+  borderRadius: '40px',
+  fontSize: '12px',
+  fontWeight: 'bold',
+};
+
+export const Label = (item: {name: string}) => {
+  if (item.name.endsWith('(deprecated)')) {
+    return (
+      <div style={flexStyles}>
+        <span>{item.name.replace('(deprecated)', '')}</span>
+        <span style={deprecatedBoxStyles}>Deprecated</span>
+      </div>
+    );
+  }
+
+  return name;
+};