+++ packages/canvas-tokens-docs/.storybook/main.ts
@@ -5,6 +5,7 @@ import {mergeConfig} from 'vite';
 const config: StorybookConfig = {
   stories: ['../**/*.stories.@(mdx|tsx)', '../stories/*.mdx'],
   addons: ['@storybook/addon-essentials'],
+
   framework: {
     name: '@storybook/react-vite',
     options: {},