+++ packages/canvas-tokens-docs/stories/base/visual-testing/DeprecatedBase.stories.tsx
@@ -2,7 +2,7 @@ import {DeprecatedColorTokens} from '../examples/Color';
 
 export default {
   component: DeprecatedColorTokens,
-  title: 'Visual Tests/Deprecated Base Tokens',
+  title: 'Docs/Old Base Tokens (deprecated)',
 };
 
 export const DeprecatedColors = {