+++ packages/canvas-tokens/utils/transformers/index.ts
@@ -60,11 +60,12 @@ export const transforms: Record<string, Transform> = {
     matcher: filter.isSysColor,
     transformer: token => {
       // eslint-disable-next-line no-useless-escape
-      return token.original.value.replace(/{[\w\.]*}/g, (a: string) => {
+      const updatedValue = token.original.value.replace(/{[\w\.]*}/g, (a: string) => {
         const cssVar = `var(--cnvs-${a.replace(/{|}/g, '').replace(/\./g, '-')})`;
-
         return cssVar.includes('palette') ? `from ${cssVar} l c h ` : ' ' + cssVar;
       });
+
+      return token.name.includes('transparent') ? 'transparent' : updatedValue;
     },
   },
   'value/opacity': {